FISH-6034 Rename package and module
