+++ deployment-transformer/deployment-transformer-api/pom.xml
@@ -46,13 +46,13 @@
     <modelVersion>4.0.0</modelVersion>
 
 	<parent>
-		<groupId>fish.payara.eclipse.transformer</groupId>
-		<artifactId>org.eclipse.transformer.payara.deployment.parent</artifactId>
+		<groupId>fish.payara.deployment.transformer</groupId>
+		<artifactId>deployment-transformer-parent</artifactId>
 		<version>1.0-SNAPSHOT</version>
 	</parent>
 
-    <artifactId>org.eclipse.transformer.payara.deployment.api</artifactId>
-	<version>${org.eclipse.transformer.payara.deployment.api.version}</version>
+    <artifactId>deployment-transformer-api</artifactId>
+	<version>${deployment.transformer.api.version}</version>
 
 	<packaging>glassfish-jar</packaging>
