+++ deployment-transformer/deployment-transformer-api/src/main/java/fish/payara/deployment/transformer/JakartaNamespaceDeploymentTransformer.java
@@ -37,7 +37,7 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-package fish.payara.eclipse.transformer.payara.deployment;
+package fish.payara.deployment.transformer;
 
 import org.glassfish.api.admin.AdminCommandContext;
 import org.glassfish.hk2.classmodel.reflect.Types;
