+++ deployment-transformer/deployment-transformer-impl/pom.xml
@@ -46,13 +46,13 @@
     <modelVersion>4.0.0</modelVersion>
 
 	<parent>
-		<groupId>fish.payara.eclipse.transformer</groupId>
-		<artifactId>org.eclipse.transformer.payara.deployment.parent</artifactId>
+		<groupId>fish.payara.deployment.transformer</groupId>
+		<artifactId>deployment-transformer-parent</artifactId>
 		<version>1.0-SNAPSHOT</version>
 	</parent>
 
-	<artifactId>org.eclipse.transformer.payara.deployment.impl</artifactId>
-	<version>${org.eclipse.transformer.payara.deployment.impl.version}</version>
+	<artifactId>deployment-transformer-impl</artifactId>
+	<version>${deployment.transformer.impl.version}</version>
 
 	<packaging>glassfish-jar</packaging>
 
@@ -65,8 +65,8 @@
 			<artifactId>internal-api</artifactId>
 		</dependency>
 		<dependency>
-			<groupId>fish.payara.eclipse.transformer</groupId>
-			<artifactId>org.eclipse.transformer.payara.deployment.api</artifactId>
+			<groupId>fish.payara.deployment.transformer</groupId>
+			<artifactId>deployment-transformer-api</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.transformer</groupId>
