+++ deployment-transformer/deployment-transformer-impl/src/main/resources/META-INF/services/fish.payara.deployment.transformer.JakartaNamespaceDeploymentTransformer
@@ -1,0 +1,1 @@
+fish.payara.deployment.transformer.JakartaNamespaceDeploymentTransformerImpl
