--- org.eclipse.transformer.payara.deployment/org.eclipse.transformer.payara.deployment.impl/src/main/resources/META-INF/services/fish.payara.eclipse.transformer.payara.deployment.JakartaNamespaceDeploymentTransformer
@@ -1,1 +1,0 @@
-fish.payara.eclipse.transformer.payara.deployment.JakartaNamespaceDeploymentTransformerImpl
